pyproject.toml
[tool.poetry]
name = "data-mining-apriori"
version = "0.1.0"
description = ""
authors = ["pfenningpat <pat.pfenning@blastpoint.com>"]
readme = "README.md"
packages = [{include = "apriori"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
mlxtend = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
